+++ helper/random/registry.go
