+++ helper/random/registry_test.go
